photodb_yolo_object_detection/requirements.txt
setuptools
PyYAML
pandas
torch>=2.3.1
ultralytics>=8.3.28